pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "My AI Projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.0.0",
    "kagglehub>=0.3.13",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]
